refactor: LottoNumbers 기능 Lotto 클래스로 이동
